+++ kicost/distributors/api_octopart.py
@@ -252,8 +252,17 @@ class api_octopart(distributor_class):
                                 # A better alternative may be to examine the packaging field of the offer.
                                 part_qty_increment = float("inf")
 
-                            # Use the qty increment to select the part SKU, web page, and available quantity.
-                            # Do this if this is the first part offer from this dist.
+                             # Select the part SKU, web page, and available quantity.
+                            # Each distributor can have differente stock codes for the same part in different
+                            # quantities / delivery package styles: cut-tape, reel, ...
+                            # Therefore we select and overwrite a previous selection if one of the
+                            # following conditions is met:
+                            #   1. We don't have a selection for this part from this ditributor yet.
+                            #   2. The MOQ is smaller than for the current selection.
+                            #   3. The part_qty_increment for this offer smaller than that of the existing selection.
+                            #      (we prefer cut-tape style packaging over reels)
+                            #   4. For DigiKey, we can't use part_qty_increment to distinguish between
+                            #      reel and cut-tape, so we need to look at the actual DigiKey part number.
                             if not parts[i].part_num[dist]:
                                 if not part.qty_avail[dist] or (offer.get('in_stock_quantity') and part.qty_avail[dist]<offer.get('in_stock_quantity')):
                                     # Keeps the information of more availability.
@@ -271,12 +280,15 @@ class api_octopart(distributor_class):
                                 if not part.qty_avail[dist] or (offer.get('in_stock_quantity') and part.qty_avail[dist]<offer.get('in_stock_quantity')):
                                 # Keeps the information of more availability.
                                     parts[i].qty_avail[dist] = offer.get('in_stock_quantity')
-                                if not part.moq[dist] or (offer.get('moq') and part.moq[dist]>offer.get('moq')):
-                                    # Save the link, stock code, ... of the page for minimum purchase.
-                                    part.moq[dist] = offer.get('moq') # Minimum order qty.
-                                    parts[i].part_num[dist] = offer.get('sku')
-                                    parts[i].url[dist] = offer.get('product_url')
-                                    parts[i].qty_increment[dist] = part_qty_increment
+                                if not part.moq[dist] or (offer.get('moq') or \
+                                   (not part.moq[dist] or (offer.get('moq') and part.moq[dist]>offer.get('moq'))) or \
+                                   (part_qty_increment < part.qty_increment[dist]) or \
+                                   (dist == "digikey" and part.part_num[dist].endswith("DKR-ND") and dist_part_num.endswith("CT-ND")):
+                                        # Save the link, stock code, ... of the page for minimum purchase.
+                                        part.moq[dist] = offer.get('moq') # Minimum order qty.
+                                        parts[i].part_num[dist] = offer.get('sku')
+                                        parts[i].url[dist] = offer.get('product_url')
+                                        parts[i].qty_increment[dist] = part_qty_increment
 
                             # Don't bother with any extra info from the distributor.
                             parts[i].info_dist[dist] = {}