+++ kicost/HISTORY.rst
@@ -3,8 +3,9 @@
 History
 -------
 
-1.1.3 (2019-10-??)
+1.1.3 (2019-12-??)
 ______________________
+* Fix the Digikey using Digi-Reel information.
 * Added LCSC as distributor into the Kitspace API (the list can be used to JLCPCB).
 * Parinfo kitspace API now just return the asked distributors (before was given all the avaliable distributors).